Separated proxy into a DLL. Fixes #30.
