+++ Proxy/Diagnostics/Logger.cs
