+++ Proxy/Exceptions/InvalidConfigurationException.cs
